增加交互,获取用户信息和答案
