+++ info.py
@@ -1,0 +1,36 @@
+login_url = "https://js.lgb360.com/lgb/user/loginByPassword.do"
+
+startComp = "https://aqy-app.lgb360.com/aqy/ques/startCompetition"
+
+answerQues = "https://aqy-app.lgb360.com/aqy/ques/answerQues"
+
+submit_url = "https://aqy-app.lgb360.com/aqy/ques/submitCompetition"
+
+login_header = {
+    "token": "",
+    "memberId": "",
+    "mobileTerminal": "0",
+    "appversion": "3.0.3",
+    "User-Agent": "LGB/3.0.3 (HMA-AL00; Android 10; zh_CN_#Hans; e4802ed3-c836-4ba8-822c-51dc9cdac4f8; 3626613241)",
+    "Content-Type": "application/x-www-form-urlencoded",
+    "Host": "js.lgb360.com",
+    "Connection": "Keep-Alive",
+    "Accept-Encoding": "gzip"
+}
+
+header = {
+    "Host": "aqy-app.lgb360.com",
+    "accept": "application/json, text/plain, */*",
+    "token": "" ,
+    "user-agent": "",
+    "memberid": "",
+    "content-type": "application/json;charset=UTF-8",
+    "origin": "https://aqy-app.lgb360.com",
+    "x-requested-with": "com.hxak.liangongbao",
+    "sec-fetch-site": "same-origin",
+    "sec-fetch-mode": "cors",
+    "sec-fetch-dest": "empty",
+    "accept-encoding": "gzip, deflate",
+    "accept-language": "zh-CN,zh;q=0.9,en-US;q=0.8,en;q=0.7",
+}
+
